+++ packages/json-serializr/src/__tests__/index.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from '@jest/globals';
-import type { Result } from '@result/result';
-import { Err, Ok } from '@result/result';
+import type { Result } from '@rustresult/result';
+import { Err, Ok } from '@rustresult/result';
 import { createModelSchema, deserialize, list, object, primitive, serialize } from 'serializr';
 
 import { createResultModelSchema, resultPropSchema } from '../index';