+++ packages/json/src/__tests__/index.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from '@jest/globals';
-import { Err, Ok } from '@result/result';
+import { Err, Ok } from '@rustresult/result';
 
 import { ResultJSON } from '../index';