+++ packages/typeorm/src/__tests__/index.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, jest } from '@jest/globals';
-import { Err, ErrAsync, Ok, OkAsync } from '@result/result';
+import { Err, ErrAsync, Ok, OkAsync } from '@rustresult/result';
 import type { DataSource, EntityManager, QueryRunner } from 'typeorm';
 import type { IsolationLevel } from 'typeorm/driver/types/IsolationLevel';