+++ packages/json-serializr/CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
-## @result/json-serializr v0.6.0 (Unreleased)
+## @rustresult/json-serializr v0.6.0 (Unreleased)
 ### Features
 
-Provides `serializr` helpers to support JSON (de)serialization of `@result/result`:
+Provides `serializr` helpers to support JSON (de)serialization of `@rustresult/result`:
 - `createResultModelSchema`
 - `resultPropSchema`