+++ packages/json/CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
-## @result/json v0.6.0 (Unreleased)
+## @rustresult/json v0.6.0 (Unreleased)
 ### Features
 
-Adds a simple JSON (de)serialization support for `@result/result`:
+Adds a simple JSON (de)serialization support for `@rustresult/result`:
 - `ResultJSON.serialize`
 - `ResultJSON.deserialize`