+++ packages/typeorm/CHANGELOG.md
@@ -1,5 +1,5 @@
 # CHANGELOG
-## @result/typeorm v0.6.0 (Unreleased)
+## @rustresult/typeorm v0.6.0 (Unreleased)
 ### Features
 
 Add TypeORM helpers: